pyproject.toml
[tool.poetry]
name = "mmbackend"
version = "0.1.0"
description = ""
authors = ["Ari <me@ari.io>"]
readme = "README.md"
packages = [{ include = "*", from = "src" }, { include = "*", from = "cli" }]

[tool.poetry.scripts]
pybe = "cli:cli"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.19"
psutil = "^5.9.8"
pydantic-settings = "^2.3.1"
click = "^8.1.7"
rich = "^13.7.1"
loguru = "^0.7.2"
pycryptodome = "^3.20.0"
pyjwt = "^2.8.0"
asyncpg = "^0.29.0"
boto3 = "^1.34.122"
botocore = "^1.34.122"
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
pytest-watch = "^4.2.0"
pyclean = "^3.0.0"
black = "^24.4.2"
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"
httpx = "^0.27.0"
pytest = "^8.2.2"
aiosqlite = "^0.20.0"
greenlet = "^3.0.3"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
count = true

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = "tests"
